+++ CHANGELOG.md
@@ -17,6 +17,7 @@ UNRELEASED
 * Fixed: use of REACT_APP_USE_LOCAL_DATA for local development
 * Changed: to show notification when there is an senscalc error after mapping target-observation
 * Changed: Added validation to target fields 
+* Added: Calibration page content
 
 
 8.0.0